Update server.js
Implemented the GET Route parameters method for the path /api/exercise/users to view all the users in the DB and for the path /api/exercise/log/_id to view user's document with a log for total exercises added and total exercise count. Expanded the functionality of the  POST method for the path /api/exercise/add so every user now has an array with all the exercise entries added based on his _id. For this to work I had nest a new mongoose Schema, for the exercise entries, in the original userSchema and make the necessary code changes in both POST methods. 4 of 5 user stories now working and for the last one I have to write another GET method for retrieving only a specific part of a user's log based on optional parameters of from & to or limit.