Update server.js
Implemented a series of new checks and error handling using the Error object. All information entered via the input fields has new strict requirements and the forms are not submitted (posted) if not all conditions pass. I started to work on the User Story #5 and for the next commit I think it will be done.