Create api.js
Added the api.js routes module to the project's repo. Implemented the final "User Story" #5, where users can get a complete or partial log with the user's exercises. Optional are "from" and "to" dates and the "limit" for viewable results. Also added a number of new conditions to avoid error while adding info to the forms. Optimized the project by using modules to breakdown the code into more readable parts and based on Node.js best practices examples. The project is now final, fully functional and ready to be submitted to FCC for review.